refactor: 로직 커스텀 훅 분리
